Remove subcommands and move agent funcs to root commands.
